Fixing weird problems with mouse controls. Changing projection matrix configuration to use gluPerspective and allow very close model viewing.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/utils@195 720011fe-810a-0410-ac5b-998ee3860a72

+++ gl_visualizer/sandbox/gui/gl_input_manager.h
@@ -43,7 +43,7 @@ class gl_input_manager
     button_state m_middle;     ///< The left-button state.
     button_state m_right;      ///< The left-button state.
 
-    int m_deadzone{5};         ///< Ignore drags shorter than this many pixels.
+    size_t m_deadzone{5};      ///< Ignore drags shorter than this many pixels.
     float m_sensitivity{.1};   ///< Higher values make the mouse more sensitive.
 
     ///@}
